caesar-project-reorganize add new structure of the project, update ignore and readme
