+++ .gitignore
@@ -1,2 +1,7 @@
 bin
 .idea
+.DS_Store
+out
+
+input*.txt
+input_ru*.txt
