+++ src/main/java/com/dudarev/caesar/enums/Languages.java
@@ -1,0 +1,5 @@
+package com.dudarev.caesar.enums;
+
+public enum Languages {
+    EN, RU
+}
