+++ src/main/java/com/dudarev/caesar/enums/Operations.java
@@ -1,0 +1,5 @@
+package com.dudarev.caesar.enums;
+
+public enum Operations {
+    decode, encode
+}
